+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `http` and `grpc` `ServerOptions` to wrap a slice of options into a single element
+
 
 ## 0.1.0 - 2019-09-25